fix compiler warning
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@17541 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/frontends/qt4/QCitation.C
@@ -45,7 +45,7 @@ template<typename String> static QStringList to_qstring_list(vector<String> cons
 vector<string> const to_string_vector(QStringList const & qlist)
 {
 	vector<string> v;
-	for (size_t i=0; i != qlist.size(); ++i) {
+	for (int i = 0; i != qlist.size(); ++i) {
 		if (qlist[i].isEmpty())
 			continue;
 		v.push_back(lyx::fromqstr(qlist[i]));
